% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHD_package.R
\docType{package}
\name{AHD-package}
\alias{AHD-package}
\title{AHD: Adaptive Hellinger Distance for Testing Independence}
\description{
Implements the Adaptive Hellinger Distance (AHD) index and perform a permutation test for testing independence between two random vectors. It includes functions for computing the AHD statistic, conducting permutation test, and combining p-values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JiujingWu/AHD}
  \item Report bugs at \url{https://github.com/JiujingWu/AHD/issues}
}

}
\author{
\strong{Maintainer}: Jiujing Wu \email{jiujingwu@foxmail.com} [copyright holder]

Authors:
\itemize{
  \item Wenwen Guo \email{guowenwen114@163.com}
}

}
\keyword{internal}
