% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHD_functions.R
\name{AHD}
\alias{AHD}
\title{Adaptive Hellinger Distance (AHD)}
\usage{
AHD(x, y, method = "manhattan")
}
\arguments{
\item{x}{A matrix containing the first random vectors.}

\item{y}{A matrix containing the second random vectors.}

\item{method}{The method for calculating the distance between two random vectors.
This must be one of \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
Any unambiguous substring can be given. Defaults to \code{"manhattan"}.}
}
\value{
A list of the objects about the AHD:
\itemize{
\item \code{AHD}: The computed \eqn{T} value;
\item \code{AHD11}: The computed \eqn{T_{11}} value;
\item \code{AHD12}: The computed \eqn{T_{12}} value;
\item \code{AHD21}: The computed \eqn{T_{21}} value;
\item \code{AHD22}: The computed \eqn{T_{22}} value.
}
}
\description{
This function computes the AHD between two random vectors.
}
