% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHD_functions.R
\name{AHD_test_4bins}
\alias{AHD_test_4bins}
\title{Adaptive Hellinger Distance (AHD) independence test}
\usage{
AHD_test_4bins(x, y, method, B = 299)
}
\arguments{
\item{x}{A matrix containing the first random vectors.}

\item{y}{A matrix containing the second random vectors.}

\item{method}{The method for calculating the distance between two random vectors.
This must be one of \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
Any unambiguous substring can be given. Defaults to \code{"manhattan"}.}
}
\value{
A list of the objects about the p-value:
\itemize{
\item \code{pvalue}: The p-value computed from the test statistic \eqn{T};
\item \code{pvalue11}: The p-value computed from the test statistic \eqn{T_{11}};
\item \code{pvalue12}: The p-value computed from the test statistic \eqn{T_{12}};
\item \code{pvalue21}: The p-value computed from the test statistic \eqn{T_{21}};
\item \code{pvalue22}: The p-value computed from the test statistic \eqn{T_{22}}.
}
}
\description{
Implement the AHD independence test via permutation test.
}
