% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOV.R
\name{arccos}
\alias{arccos}
\title{Pairwise angular-distance matrix around a reference row}
\usage{
arccos(A, r)
}
\arguments{
\item{A}{A numeric matrix of size \eqn{n \times p} with sample size \eqn{n} and dimension \eqn{p}.}

\item{r}{An integer in \code{1:n} specifying the reference row about which the row-wise centering is performed.}
}
\value{
A numeric matrix of angular distances.
}
\description{
Computes the pairwise angular-distance matrix (in radians) for rows of \code{A} centered at a reference row.
}
\seealso{
\link[=pcov2]{pcov2}, \link[=pcov2_test]{pcov2_test}
}
