% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHD_functions.R
\name{P_VALUE_COMB}
\alias{P_VALUE_COMB}
\title{Adaptive Hellinger Distance (AHD) independence test}
\usage{
P_VALUE_COMB(pvalue11, pvalue12, pvalue21, pvalue22)
}
\arguments{
\item{pvalue11}{The p-value computed from the test statistic \eqn{T_{11}}.}

\item{pvalue12}{The p-value computed from the test statistic \eqn{T_{12}}.}

\item{pvalue21}{The p-value computed from the test statistic \eqn{T_{21}}.}

\item{pvalue22}{The p-value computed from the test statistic \eqn{T_{22}}.}
}
\value{
A list of the objects about the combined p-value:
\itemize{
\item \code{BH}: Combined p-values are adjusted via the Benjamini–Hochberg (BH) method;
\item \code{CCT}: Combined p-values are adjusted via the Cauchy Combination Test (CCT) method.
}
}
\description{
Implement the AHD independence test via the combined p-values.
}
