% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOV.R
\name{pcov2}
\alias{pcov2}
\title{Projection Covariance}
\usage{
pcov2(x, y)
}
\arguments{
\item{x}{A numeric matrix of size n * p.}

\item{y}{A numeric matrix of size n * q.}
}
\value{
A list with two elements:
\item{pcov}{The projection covariance of x and y.}
\item{S2}{A component of the projection covariance, which is required for calculating the subsequent test statistic.}
}
\description{
Calculate the projection covariance of two random vectors. Two vectors can be with different dimensions, but with equal sample sizes.
}
\references{
L.Zhu, K.Xu, R.Li, W.Zhong (2017). Projection correlation between two random vectors. Biometrika, Volume 104, Pages 829-843. https://doi.org/10.1093/biomet/asx043
}
