% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCOV.R
\name{pcov2_test}
\alias{pcov2_test}
\title{Projection Correlatoin Permutation Test}
\usage{
pcov2_test(x, y, B = 299)
}
\arguments{
\item{x}{A numeric matrix of size n * p.}

\item{y}{A numeric matrix of size n * q.}

\item{B}{Integer, the number of permutations. Defaults to \code{299}.}
}
\value{
A list with:
\item{stat}{The observed test statistic \code{stat}.}
\item{pvalue}{The permutation p-value \code{pvalue}.}
}
\description{
Test whether two vectors are independent. Calculate the test result of projection correlation test.
}
\references{
L.Zhu, K.Xu, R.Li, W.Zhong (2017). Projection correlation between two random vectors. Biometrika, Volume 104, Pages 829-843. https://doi.org/10.1093/biomet/asx043
}
